Inline part of the TaskBuilder hierarchy.
