--- src/ispd/motor/carga/task/TaskBuilder.java
@@ -1,24 +1,0 @@
-package ispd.motor.carga.task;
-
-import ispd.motor.filas.RedeDeFilas;
-import ispd.motor.filas.Tarefa;
-import ispd.motor.filas.servidores.CS_Processamento;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-public abstract class TaskBuilder {
-    public List<Tarefa> makeTasksEvenlyDistributedBetweenMasters(
-            final RedeDeFilas qn, final int taskCount) {
-        final var masters = qn.getMestres();
-
-        return IntStream.range(0, taskCount)
-                .map(i -> i % masters.size())
-                .mapToObj(masters::get)
-                .map(this::makeTaskFor)
-                .collect(Collectors.toList());
-    }
-
-    public abstract Tarefa makeTaskFor(final CS_Processamento master);
-}
