+++ src/ispd/motor/carga/task/TraceTaskBuilder.java
@@ -1,17 +1,21 @@
 package ispd.motor.carga.task;
 
 import ispd.motor.carga.workload.WorkloadGenerator;
+import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
 
-public class TraceTaskBuilder extends TaskBuilder {
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+public class TraceTaskBuilder {
     protected final TaskInfo taskInfo;
 
     public TraceTaskBuilder(final TaskInfo taskInfo) {
         this.taskInfo = taskInfo;
     }
 
-    @Override
     public Tarefa makeTaskFor(final CS_Processamento master) {
         return new Tarefa(
                 this.taskInfo.id(),
@@ -31,5 +35,16 @@ public class TraceTaskBuilder extends TaskBuilder {
 
     protected double calculateProcessingTime() {
         return this.taskInfo.processingTime();
+    }
+
+    public List<Tarefa> makeTasksEvenlyDistributedBetweenMasters(
+            final RedeDeFilas qn, final int taskCount) {
+        final var masters = qn.getMestres();
+
+        return IntStream.range(0, taskCount)
+                .map(i -> i % masters.size())
+                .mapToObj(masters::get)
+                .map(this::makeTaskFor)
+                .collect(Collectors.toList());
     }
 }
